requirements.txt
requests
pyzabbix
python-dotenv
pyasn1==0.4.8
pysnmp
psycopg2-binary
prettytable
